+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- Nothing yet
+
+## [0.37.0] - 2023-08-14
+
 ### Added
 
 - Pointer to stpp sample entry in StsdBox
@@ -416,7 +420,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - New unique repo name: `mp4ff`
 
-[Unreleased]: https://github.com/Eyevinn/mp4ff/compare/v0.36.0...HEAD
+[Unreleased]: https://github.com/Eyevinn/mp4ff/compare/v0.37.0...HEAD
+[0.37.0]: https://github.com/Eyevinn/mp4ff/compare/v0.36.0...v0.37.0
 [0.36.0]: https://github.com/Eyevinn/mp4ff/compare/v0.35.0...v0.36.0
 [0.35.0]: https://github.com/Eyevinn/mp4ff/compare/v0.34.1...v0.35.0
 [0.34.1]: https://github.com/Eyevinn/mp4ff/compare/v0.34.0...v0.34.1